Restructure derive
